+++ apps/shell/shell-memdebug.c
@@ -71,7 +71,7 @@ PROCESS_THREAD(shell_poke_process, ev, data)
     PROCESS_EXIT();
   }
   
-  address = (uint8_t *)(int)shell_strtolong(args, &next);
+  address = (uint8_t *)(uintptr_t)shell_strtolong(args, &next);
   if(next == args) {
     shell_output_str(&poke_command, "usage 1", "");
     PROCESS_EXIT();
@@ -106,7 +106,7 @@ PROCESS_THREAD(shell_peek_process, ev, data)
     PROCESS_EXIT();
   }
   
-  address = (uint8_t *)(int)shell_strtolong(args, &next);
+  address = (uint8_t *)(uintptr_t)shell_strtolong(args, &next);
   if(next == args) {
     shell_output_str(&peek_command, "usage 1", "");
     PROCESS_EXIT();
